Moves category finder to dedicated folder
This commit refactors the project structure by moving the `ParentCategoryFinder` service to a dedicated `finders` directory.

This change improves project organization and maintainability by grouping related finder services together.
