+++ src/modules/categories/finders/parent-category.finder.ts
@@ -1,7 +1,7 @@
 // parent-category.finder.ts (nuevo servicio especializado)
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Category } from './entities/category.entity';
+import { Category } from '../entities/category.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
